Feed parsing fixes, UI tweaks,  and a DB tweak.
Turns out doing a where clause with all the existing fields doesn't
actually match in sqlite, for some reason.  Doing it just on the key
fields works fine, so I added the creation of the key index to the
code.  It isn't persisted yet, so I have to re-do it each time
the web server starts, for now.

There are a bunch of small UI tweaks, and a number of checks and
fallbacks added for the various different real RSS feeds I'm
reading.  So much for standards, eh?

Oh, and I changed the way the DBPATH is calculated, which will
let me run the tests without trashing the "real" database, but
I haven't actually taken advantage of that fact yet.  It still
ought to go in a config file.