Add static handler and js based swipe support.
Now swipe left is next, and swipe right is prev.  I got the javascript,
which I only partially understand, from here:

    http://padilicious.com/code/touchevents/

The static content handler is based on the 'serve' script in the Python
standard library scripts directory.

I also fixed the weird placement of the 'refresh' link; I'd forgotten to
remove the float:left declaration, which resulted in it being rendered
to far to the right.  (Yeah, go figure...it's CSS.)