Restore 'last article' check and add missing ;s.
I think the idea that the last article check was slowing things down
was an illusion.  It seems to be fast enough at the moment, at
least.  Note that the article lookup is done multiple times as it
stands.  Caching would be a good idea, at some point.

Also add a mention of the swipe feature to the README.